+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changes to the Mapbox Navigation SDK for iOS
 
 ## master
+
+* This library now requires a minimum deployment target of iOS 10.0 or above. iOS 9._x_ is no longer supported. ([#2206](https://github.com/mapbox/mapbox-navigation-ios/pull/2206))
 * Fixed an issue where manually incrementing `RouteProgress.legIndex` could lead to undefined behavior. ([#2229](https://github.com/mapbox/mapbox-navigation-ios/pull/2229))
 
 ## v0.37.0